Dedup snapshot fetch code in odsp driver
