+++ packages/drivers/odsp-driver/src/fetchSnapshot.ts
@@ -3,14 +3,25 @@
  * Licensed under the MIT License.
  */
 
+import { default as AbortController } from "abort-controller";
+import { v4 as uuid } from "uuid";
 import { ITelemetryLogger } from "@fluidframework/common-definitions";
+import { assert } from "@fluidframework/common-utils";
 import { PerformanceEvent } from "@fluidframework/telemetry-utils";
-import { ISnapshotOptions }  from "@fluidframework/odsp-driver-definitions";
-import { IOdspSnapshot } from "./contracts";
+import {
+    ICacheEntry,
+    IEntry,
+    IOdspResolvedUrl,
+    IPersistedCache,
+    ISnapshotOptions,
+    snapshotKey,
+    TokenFetchOptions,
+} from "@fluidframework/odsp-driver-definitions";
+import { IOdspSnapshot, IVersionedValueWithEpoch } from "./contracts";
 import { EpochTracker } from "./epochTracker";
 import { getQueryString } from "./getQueryString";
 import { getUrlAndHeadersWithAuth } from "./getUrlAndHeadersWithAuth";
-import { IOdspResponse } from "./odspUtils";
+import { fetchAndParseAsJSONHelper, IOdspResponse, ISnapshotCacheValue } from "./odspUtils";
 
 /**
  * Fetches a snapshot from the server with a given version id.
@@ -50,4 +61,236 @@ export async function fetchSnapshot(
         },
         async () => epochTracker.fetchAndParseAsJSON<IOdspSnapshot>(url, { headers }, "snapshotTree"),
     );
+}
+
+export async function fetchLatestSnapshotCore(
+    odspResolvedUrl: IOdspResolvedUrl,
+    storageTokenFetcher: (options: TokenFetchOptions, name: string) => Promise<string | null>,
+    tokenFetchOptions: TokenFetchOptions,
+    snapshotOptions: ISnapshotOptions | undefined,
+    logger: ITelemetryLogger,
+    summarizerClient: boolean = false,
+    epochTracker: EpochTracker | undefined,
+    snapshotCacheEntry: IEntry,
+    persistedCache: IPersistedCache,
+): Promise<{
+    odspSnapshot: ISnapshotCacheValue,
+    cacheP?: Promise<void>,
+}> {
+    const snapshotUrl = odspResolvedUrl.endpoints.snapshotStorageUrl;
+    const url = `${snapshotUrl}/trees/latest?ump=1`;
+    const storageToken = await storageTokenFetcher(tokenFetchOptions, "PrefetchLatestSnapshot");
+    assert(storageToken !== null, "Storage token should not be null");
+    const formBoundary = uuid();
+    const formParams: string[] = [];
+    formParams.push(`--${formBoundary}`);
+    formParams.push(`Authorization: Bearer ${storageToken}`);
+    formParams.push(`X-HTTP-Method-Override: GET`);
+    const logOptions = {};
+    if (snapshotOptions !== undefined) {
+        Object.entries(snapshotOptions).forEach(([key, value]) => {
+            if (value !== undefined) {
+                formParams.push(`${key}: ${value}`);
+                logOptions[`snapshotOption_${key}`] = value;
+            }
+        });
+    }
+    if (odspResolvedUrl.sharingLinkToRedeem) {
+        formParams.push(`sl: ${odspResolvedUrl.sharingLinkToRedeem}`);
+    }
+    formParams.push(`_post: 1`);
+    formParams.push(`\r\n--${formBoundary}--`);
+    const postBody = formParams.join("\r\n");
+    const headers: {[index: string]: any} = {
+        "Content-Type": `multipart/form-data;boundary=${formBoundary}`,
+    };
+
+    let controller: AbortController | undefined;
+    if (summarizerClient !== true && snapshotOptions?.timeout !== undefined) {
+        controller = new AbortController();
+        setTimeout(
+            () => controller!.abort(),
+            snapshotOptions.timeout,
+        );
+    }
+
+    // This event measures only successful cases of getLatest call (no tokens, no retries).
+    return PerformanceEvent.timedExecAsync(
+        logger,
+        {
+            eventName: "TreesLatest",
+            ...logOptions,
+        },
+        async (event) => {
+            const startTime = performance.now();
+            let response: IOdspResponse<IOdspSnapshot>;
+            if (epochTracker !== undefined) {
+                response = await epochTracker.fetchAndParseAsJSON<IOdspSnapshot>(
+                    url,
+                    {
+                        body: postBody,
+                        headers,
+                        signal: controller?.signal,
+                        method: "POST",
+                    },
+                    "treesLatest",
+                    true,
+                );
+            } else {
+                response = await fetchAndParseAsJSONHelper<IOdspSnapshot>(
+                    url,
+                    {
+                        body: postBody,
+                        headers,
+                        signal: controller?.signal,
+                        method: "POST",
+                    },
+                );
+            }
+            const endTime = performance.now();
+            const overallTime = endTime - startTime;
+            const snapshot: IOdspSnapshot = response.content;
+            let dnstime: number | undefined; // domainLookupEnd - domainLookupStart
+            let redirectTime: number | undefined; // redirectEnd -redirectStart
+            let tcpHandshakeTime: number | undefined; // connectEnd  - connectStart
+            let secureConntime: number | undefined; // connectEnd  - secureConnectionStart
+            let responseTime: number | undefined; // responsEnd - responseStart
+            let fetchStToRespEndTime: number | undefined; // responseEnd  - fetchStart
+            let reqStToRespEndTime: number | undefined; // responseEnd - requestStart
+            let networkTime: number | undefined; // responseEnd - startTime
+            const spReqDuration = response.headers.get("sprequestduration");
+
+            // getEntriesByType is only available in browser performance object
+            const resources1 = performance.getEntriesByType?.("resource") ?? [];
+            // Usually the latest fetch call is to the end of resources, so we start from the end.
+            for (let i = resources1.length - 1; i > 0; i--) {
+                const indResTime = resources1[i] as PerformanceResourceTiming;
+                const resource_name = indResTime.name;
+                const resource_initiatortype = indResTime.initiatorType;
+                if ((resource_initiatortype.localeCompare("fetch") === 0) && (resource_name.localeCompare(url) === 0)) {
+                    redirectTime = indResTime.redirectEnd - indResTime.redirectStart;
+                    dnstime = indResTime.domainLookupEnd - indResTime.domainLookupStart;
+                    tcpHandshakeTime = indResTime.connectEnd - indResTime.connectStart;
+                    secureConntime = (indResTime.secureConnectionStart > 0) ?
+                        (indResTime.connectEnd - indResTime.secureConnectionStart) : 0;
+                    responseTime = indResTime.responseEnd - indResTime.responseStart;
+                    fetchStToRespEndTime = (indResTime.fetchStart > 0) ?
+                        (indResTime.responseEnd - indResTime.fetchStart) : 0;
+                    reqStToRespEndTime = (indResTime.requestStart > 0) ?
+                        (indResTime.responseEnd - indResTime.requestStart) : 0;
+                    networkTime = (indResTime.startTime > 0) ? (indResTime.responseEnd - indResTime.startTime) : 0;
+                    if (spReqDuration) {
+                        networkTime = networkTime - parseInt(spReqDuration, 10);
+                    }
+                    break;
+                }
+            }
+
+            const { numTrees, numBlobs, encodedBlobsSize, decodedBlobsSize } = evalBlobsAndTrees(snapshot);
+            const clientTime = networkTime ? overallTime - networkTime : undefined;
+
+            // There are some scenarios in ODSP where we cannot cache, trees/latest will explicitly tell us when we
+            // cannot cache using an HTTP response header.
+            const canCache = response.headers.get("disablebrowsercachingofusercontent") !== "true";
+            // There maybe no snapshot - TreesLatest would return just ops.
+            const sequenceNumber: number = (snapshot.trees && (snapshot.trees[0] as any).sequenceNumber) ?? 0;
+            const seqNumberFromOps = snapshot.ops && snapshot.ops.length > 0 ?
+                snapshot.ops[0].sequenceNumber - 1 :
+                undefined;
+
+            const value: ISnapshotCacheValue = { snapshot, sequenceNumber };
+            let cacheP: Promise<void> | undefined;
+            if (!Number.isInteger(sequenceNumber)
+                || seqNumberFromOps !== undefined && seqNumberFromOps !== sequenceNumber) {
+                logger.sendErrorEvent({ eventName: "fetchSnapshotError", sequenceNumber, seqNumberFromOps });
+                value.sequenceNumber = undefined;
+            } else if (canCache) {
+                if (epochTracker !== undefined) {
+                    cacheP = epochTracker.put(
+                        snapshotCacheEntry,
+                        value,
+                    ).catch(() => {});
+                } else {
+                    const fluidEpoch = response.headers.get("x-fluid-epoch");
+                    assert(fluidEpoch !== undefined, "Epoch  should be present in response");
+                    const valueWithEpoch: IVersionedValueWithEpoch = {
+                        value,
+                        fluidEpoch,
+                        version: 2,
+                    };
+                    const cacheEntry: ICacheEntry = {
+                        type: snapshotKey,
+                        key: "",
+                        file: {
+                            resolvedUrl: odspResolvedUrl,
+                            docId: odspResolvedUrl.hashedDocumentId,
+                        },
+                    };
+                    cacheP = persistedCache.put(
+                        cacheEntry,
+                        valueWithEpoch,
+                    ).catch(() => {});
+                }
+            }
+            event.end({
+                trees: numTrees,
+                blobs: snapshot.blobs?.length ?? 0,
+                leafNodes: numBlobs,
+                encodedBlobsSize,
+                decodedBlobsSize,
+                sequenceNumber,
+                ops: snapshot.ops?.length ?? 0,
+                headers: Object.keys(headers).length !== 0 ? true : undefined,
+                redirecttime: redirectTime,
+                dnsLookuptime: dnstime,
+                responsenetworkTime: responseTime,
+                tcphandshakeTime: tcpHandshakeTime,
+                secureconnectiontime: secureConntime,
+                fetchstarttorespendtime: fetchStToRespEndTime,
+                reqstarttorespendtime: reqStToRespEndTime,
+                overalltime: overallTime,
+                networktime: networkTime,
+                clienttime: clientTime,
+                // Sharing link telemetry regarding sharing link redeem status and performance. Ex: FRL; dur=100, FRS;
+                // desc=S, FRP; desc=False. Here, FRL is the duration taken for redeem, FRS is the redeem
+                // status (S means success), and FRP is a flag to indicate if the permission has changed.
+                sltelemetry: response.headers.get("x-fluid-sltelemetry"),
+                ...response.commonSpoHeaders,
+            });
+            return { odspSnapshot: value, cacheP };
+        },
+    ).catch((error) => {
+        // Issue #5895:
+        // If we are offline, this error is retryable. But that means that RetriableDocumentStorageService
+        // will run in circles calling getSnapshotTree, which would result in this class going getVersions / individual
+        // blob download path. This path is very slow, and will not work with delay-loaded data stores
+        // and ODSP storage deleting old snapshots and blobs.
+        if (typeof error === "object" && error !== null) {
+            error.canRetry = false;
+        }
+        throw error;
+    });
+}
+
+export function evalBlobsAndTrees(snapshot: IOdspSnapshot) {
+    let numTrees = 0;
+    let numBlobs = 0;
+    let encodedBlobsSize = 0;
+    let decodedBlobsSize = 0;
+    for (const tree of snapshot.trees) {
+        for (const treeEntry of tree.entries) {
+            if (treeEntry.type === "blob") {
+                numBlobs++;
+            } else if (treeEntry.type === "tree") {
+                numTrees++;
+            }
+        }
+    }
+    if (snapshot.blobs !== undefined) {
+        for (const blob of snapshot.blobs) {
+            decodedBlobsSize += blob.size;
+            encodedBlobsSize += blob.content.length;
+        }
+    }
+    return { numTrees, numBlobs, encodedBlobsSize, decodedBlobsSize };
 }
